Extract Kibana Migrator Test Kit into separate package
