+++ src/core/server/integration_tests/saved_objects/service/lib/update.test.ts
@@ -14,10 +14,7 @@ import type { ElasticsearchClient } from '@kbn/core-elasticsearch-server';
 import type { SavedObjectsType, SavedObjectsModelVersionMap } from '@kbn/core-saved-objects-server';
 import { type TestElasticsearchUtils } from '@kbn/core-test-helpers-kbn-server';
 import '../../migrations/jest_matchers';
-import {
-  getKibanaMigratorTestKit,
-  startElasticsearch,
-} from '../../migrations/kibana_migrator_test_kit';
+import { getKibanaMigratorTestKit, startElasticsearch } from '@kbn/migrator-test-kit';
 import { getBaseMigratorParams } from '../../migrations/fixtures/zdt_base.fixtures';
 
 export const logFilePath = Path.join(__dirname, 'update.test.log');
