+++ src/core/tsconfig.json
@@ -174,6 +174,7 @@
     "@kbn/core-pricing-server-internal",
     "@kbn/core-pricing-common",
     "@kbn/lazy-object",
+    "@kbn/migrator-test-kit",
   ],
   "exclude": [
     "target/**/*",
