+++ src/platform/packages/private/kbn-migrator-test-kit/src/elasticsearch_client_wrapper.ts
