+++ tsconfig.base.json
@@ -1394,6 +1394,8 @@
       "@kbn/metrics-data-access-plugin/*": ["x-pack/solutions/observability/plugins/metrics_data_access/*"],
       "@kbn/metrics-experience-plugin": ["src/platform/plugins/shared/metrics_experience"],
       "@kbn/metrics-experience-plugin/*": ["src/platform/plugins/shared/metrics_experience/*"],
+      "@kbn/migrator-test-kit": ["src/platform/packages/private/kbn-migrator-test-kit"],
+      "@kbn/migrator-test-kit/*": ["src/platform/packages/private/kbn-migrator-test-kit/*"],
       "@kbn/ml-agg-utils": ["x-pack/platform/packages/private/ml/agg_utils"],
       "@kbn/ml-agg-utils/*": ["x-pack/platform/packages/private/ml/agg_utils/*"],
       "@kbn/ml-anomaly-utils": ["x-pack/platform/packages/shared/ml/anomaly_utils"],
